multiple improvements to lesson printing
