+++ src/main/java/com/lorepo/icplayer/client/module/sourcelist/SourceListModule.java
@@ -470,16 +470,8 @@ public class SourceListModule extends BasicModuleModel implements IWCAGModuleMod
 
 	@Override
 	public String getPrintableHTML(boolean showAnswers) {
-		if (getPrintable() == PrintableMode.NO) return null;
-		
-		String result = "<div class=\"ic_sourceList\" id=\"" + getId() +"\">";
-		for (int i = 0; i < items.size(); i++) {
-			result += items.get(i);
-			if (i + 1 < items.size()) {
-				result += ", ";
-			}
-		}
-		result += "</div>";
+		SourceListPrintable printable = new SourceListPrintable(this);
+		String result = printable.getPrintableHTML(showAnswers);
 		return result;
 	}
