+++ src/main/java/com/lorepo/icplayer/public/libs/dictionaries/com/dictionary.js
@@ -2049,5 +2049,6 @@ var ice_dictionary_en = {
 	"Cross_Lesson_property_page": "Page",
 	"Cross_Lesson_property_type": "Type",
 	"printable_name_label": "Printable",
-	"printable_is_section": "is section"
+	"printable_is_section": "is section",
+	"printable_preview": "Printable preview"
 };
