+++ src/main/java/com/lorepo/icplayer/public/libs/dictionaries/fr/dictionary.js
@@ -2028,5 +2028,6 @@ var ice_dictionary_fr = {
 	"Cross_Lesson_property_page": "Page",
 	"Cross_Lesson_property_type": "Type",
 	"printable_name_label": "Printable",
-	"printable_is_section": "is section"
+	"printable_is_section": "is section",
+	"printable_preview": "Printable preview"
 };
