+++ src/main/java/com/lorepo/icplayer/public/libs/dictionaries/pl/dictionary.js
@@ -2027,5 +2027,6 @@ var ice_dictionary_pl = {
 	"Cross_Lesson_property_page": "Page",
 	"Cross_Lesson_property_type": "Type",
 	"printable_name_label": "Printable",
-	"printable_is_section": "is section"
+	"printable_is_section": "is section",
+	"printable_preview": "Printable preview"
 };
